requirements.txt
pymongo~=4.3.3
fastapi~=0.94.1
starlette~=0.26.1
sqlmodel~=0.0.8
dnspython~=2.3.0
idna~=3.3
requests~=2.26.0
h11~=0.14.0
pip~=21.3.1
attrs~=21.2.0
tornado~=6.1
Jinja2~=3.0.3
dill~=0.3.6
PyYAML~=6.0
anyio~=3.4.0
sniffio~=1.2.0
click~=8.1.3
Pygments~=2.10.0
certifi~=2021.10.8
isort~=5.11.5
tomli~=2.0.1
colorama~=0.4.4
setuptools~=39.0.1
ipython~=7.30.1
MarkupSafe~=2.0.1
pylint~=2.13.9
astroid~=2.11.7
mccabe~=0.7.0
platformdirs~=2.4.1
wrapt~=1.15.0
pydantic~=1.10.6
websockets~=10.4
ujson~=5.1.0
uvicorn~=0.21.0
httptools~=0.5.0
watchfiles~=0.18.1
decorator~=5.1.0
zipp~=3.6.0
python-dotenv~=0.21.1
log4python~=1.0.3